+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 *Please add new entries at the top.*
 
 1. Fix minimum deployment target of iOS 11 in CocoaPods
+1. Fix CI release git tag push trigger (#869, kudos to @p4checo)
 
 # 7.1.1
 1. Bumped deployment target to iOS 11, tvOS 11, watchOS 4, macOS 10.13, per Xcode 14 warnings (#865, kudos to @lickel)